% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiCalc.R
\name{aiCalc}
\alias{aiCalc}
\title{Forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent}
\usage{
aiCalc(json)
}
\arguments{
\item{json}{parameters data from api.ai}
}
\value{
list
}
\description{
Todo: use cron to pre-generate model parameter estimations, always use latest estimations as start values
}

