% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiFX.R
\name{aiFX}
\alias{aiFX}
\title{Return forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent}
\usage{
aiFX(result, ...)
}
\arguments{
\item{result}{list of data from api.ai}
}
\value{
list
}
\description{
Return forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent
}

