% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerOne.R
\name{eulerOne}
\alias{eulerOne}
\title{A univariate euler simulation function for Chan et al. 1992 Stochastic Differential Equation}
\usage{
eulerOne(xinit = 1, model = "you3", Terminal = 1, nsim = 10,
  parameter = list(p1 = -0.04, p2 = 0.1, p3 = 0.23))
}
\arguments{
\item{xinit}{initial value vector of state variables (default 1, number)}

\item{model}{sde model (you3, yckls, yhdf for yuima)}

\item{Terminal}{(default 1 (year), positive number)}

\item{nsim}{number of simulations (default 10, integer)}

\item{parameter}{named list of parameters}
}
\value{
simulations
}
\description{
It can be updated to take any kind of SDE model
Simulates only at the (future) Terminal (trading time) point
}

