% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Gets latest available (our sources) univariate financial market data}
\usage{
getData(request = list(class = "FX", base_currency = "EUR", currency = "USD",
  frequency = "day", limit = 100))
}
\arguments{
\item{request}{list()}
}
\value{
vector of double (numeric)
}
\description{
FX: Always expressed as the value in base currency (this may not follow trading convention)
}

