% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlefilter.R
\name{mlefilter}
\alias{mlefilter}
\title{Filter the summary of mle result for use in javascript
This is a simple test function}
\usage{
mlefilter(mle)
}
\arguments{
\item{mle}{is the return value from a (yuima) qmle or mle estimation}
}
\value{
json of results from summary(mle)
}
\description{
Filter the summary of mle result for use in javascript
This is a simple test function
}

