% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_all.R
\name{simulate_all}
\alias{simulate_all}
\title{Estimates SDE parameters for a univariate time series, and simulates future values for this SDE}
\usage{
simulate_all(data, T = 1, nsim = 100, delta = 1/252)
}
\arguments{
\item{data}{univariate time series of type numeric}

\item{T}{risk time horizon in years}

\item{nsim}{number of simulations at time T}

\item{delta}{data unit distance}
}
\value{
json
}
\description{
Daily data for longer than 7 days risk time horizon, else hourly data (up to 7 days risk time horison)
}

