% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.sql.R
\name{sqlite.sql}
\alias{sqlite.sql}
\title{sqlite sql}
\usage{
sqlite.sql(url, stmt)
}
\arguments{
\item{url}{sqlite db path e.g. "/abs_path/my.db"}

\item{stmt}{sql statement e.g. "SELECT * FROM stdin LIMIT 100"}
}
\value{
tabular data
}
\description{
Connects to db, makes a query to db, closes connection to db
}
\details{
Single quotes are used to indicate the beginning and end of a string in SQL
Double quotes are preferred in R
synchronous: c("off", "normal", "full")

To retrieve results a chunk at a time, use dbSendQuery, dbFetch, then ClearResult. Alternatively, if you want all the results (and they'll fit in memory) use dbGetQuery which sends, fetches and clears for you.
}

