% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yuima.qmle.R
\name{yuima.qmle}
\alias{yuima.qmle}
\title{An interface to the yuima::qmle function}
\usage{
yuima.qmle(data, delta = 1/252, summary = TRUE, drift, diffusion,
  hurst = 0.5, solve.variable = "x", start, ...)
}
\arguments{
\item{data}{list(x=data) or list(x1=data1, x2=data2,...) possibly use log or return of data.}

\item{delta}{1 divided by the # of observations per year e.g. with business daily data typically: 1/252}

\item{summary}{TRUE or FALSE: TRUE}

\item{drift}{"mu * x" or c("mu1 * x1",...)}

\item{diffusion}{"sigma * x" or matrix(c("sigma1","0","0","sigma2"),2,2)}

\item{hurst}{0.5}

\item{solve.variable}{"x" or c("x1","x2")}

\item{start}{list(mu = 0.1, sigma = 0.1)}
}
\value{
maximum likelihood estimation
}
\description{
Univariate and multivariate models, example given for univariate GBM
Optional paramters: method, lower, upper
}

