% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yuima.sde.R
\name{yuima.sde}
\alias{yuima.sde}
\title{yuima.sde returns parameters for some specific (yuima) sde models}
\usage{
yuima.sde(model = "you3", multivariate = FALSE, correlated = FALSE, data)
}
\arguments{
\item{model}{(you3, yckls, yhdf_b2)}

\item{multivariate}{boolean}

\item{correlated}{boolean, calculates correlations of data if multivariate}

\item{data}{multivariate data time series for multivariate modelling}
}
\value{
list
}
\description{
Warning: dependency in eulerOne()
ToDo: add heston model
}

