% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yuima.simulate.R
\name{yuima.simulate}
\alias{yuima.simulate}
\title{An interface to the yuima::simulate function}
\usage{
yuima.simulate(setseed = FALSE, sumsim = TRUE, nsim = 1, drift, diffusion,
  hurst = 0.5, solve.variable = "x", xinit = 1, Terminal = 1, n = 100,
  parameter, grid = NULL)
}
\arguments{
\item{setseed}{boolean (default FALSE)}

\item{sumsim}{sum simulations vertically (default TRUE)}

\item{nsim}{number of simulations (default 1)}

\item{drift}{(e.g."mu * x")}

\item{diffusion}{(e.g. "sigma * x")}

\item{hurst}{0.5}

\item{solve.variable}{"x"  (state.variable is set equal to this)}

\item{xinit}{initial value vector of state variables (default 1)}

\item{Terminal}{(default 1)}

\item{n}{number of trading times (default 1)}

\item{parameter}{named list of parameters (name true.parameter in yuima)}

\item{grid}{for setSampling}
}
\value{
simulations
}
\description{
Multivariate models
}

