% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiUpdown.R
\name{aiUpdown}
\alias{aiUpdown}
\title{Forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent}
\usage{
aiUpdown(result, ...)
}
\arguments{
\item{result}{list of data from api.ai}
}
\value{
list
}
\description{
Forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent
}

