% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiUpdown.R
\name{aiUpdown}
\alias{aiUpdown}
\title{Forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent}
\usage{

  aiUpdown(json_obj = "{\\"result\\": {\\"parameters\\": {\\"activity_financial\\": \\"import\\", \\"amount-currency\\": {\\"amount\\": 100, \\"currency\\": \\"GBP\\"}, \\"base-currency\\": \\"DKK\\", \\"date\\": \\"2017-12-10\\"}}}")
}
\arguments{
\item{json_obj}{json webhook object from api.ai}
}
\value{
list
}
\description{
Forecast of chance of profit and risk of loss for api.ai (this) updown financial activity agent
}

