% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_all.R
\name{simulate_all}
\alias{simulate_all}
\title{Estimates SDE parameters for a univariate time series, and simulates future values for this SDE}
\usage{
simulate_all(data, T = 1, nsim = 100, delta = 1/252, estimations = 3,
  neg_values = FALSE, multi_model = TRUE)
}
\arguments{
\item{data}{univariate time series of type numeric}

\item{T}{risk time horizon in years (default 1 year)}

\item{nsim}{number of simulations at time T (default 100)}

\item{delta}{data unit distance (default 1/252)}

\item{neg_values}{in data, boolean (default FALSE)}
}
\value{
vector of double (numeric)
}
\description{
Estimates SDE parameters for a univariate time series, and simulates future values for this SDE
}

